fixes according to PR review
