--- web-app/src/lib/db/dto/AgentDTO.ts
@@ -1,139 +1,0 @@
-import { Prisma } from "@prisma/client";
-import { getTranslations } from "next-intl/server";
-
-import { getEnvSecrets } from "@/config/env.config";
-import { ipfsUrlResolver } from "@/lib/ipfs";
-
-type AgentWithRelations = Prisma.AgentGetPayload<{
-  include: {
-    exampleOutput: true;
-    overrideExampleOutput: true;
-    pricing: {
-      include: {
-        fixedPricing: {
-          include: {
-            amounts: true;
-          };
-        };
-      };
-    };
-    tags: true;
-    overrideTags: true;
-    rating: true;
-  };
-}>;
-
-export interface ExampleOutputDTO {
-  readonly id: string;
-  readonly name: string;
-  readonly mimeType: string;
-}
-
-export interface LegalDTO {
-  readonly privacyPolicy: string | null;
-  readonly terms: string | null;
-  readonly other: string | null;
-}
-
-export interface AgentDTO {
-  readonly id: string;
-  readonly onChainIdentifier: string;
-  readonly name: string;
-  readonly description: string;
-  readonly exampleOutput: ExampleOutputDTO[];
-  readonly author: string;
-  readonly legal?: LegalDTO;
-  readonly tags: string[];
-  readonly image: string;
-  readonly averageStars: number | null;
-  readonly credits: number;
-  readonly apiBaseUrl: URL;
-}
-
-function calculateCredits(amount: number): number {
-  return amount;
-}
-
-export async function createAgentDTO(
-  agent: AgentWithRelations,
-): Promise<AgentDTO> {
-  if (!agent.rating || !agent.pricing.fixedPricing) {
-    throw new Error("Agent must have Rating and FixedPricing");
-  }
-
-  const t = await getTranslations("Data.Agent");
-
-  //Validate the API base URL
-  const blacklistedHostnames = getEnvSecrets().BLACKLISTED_AGENT_HOSTNAMES;
-  const apiBaseUrl = new URL(agent.apiBaseUrl);
-  if (blacklistedHostnames.includes(apiBaseUrl.hostname)) {
-    throw new Error("Agent API base URL is not allowed");
-  }
-  if (apiBaseUrl.protocol !== "https:" && apiBaseUrl.protocol !== "http:") {
-    throw new Error("Agent API base URL must be HTTP or HTTPS");
-  }
-  if (
-    apiBaseUrl.port !== "80" &&
-    apiBaseUrl.port !== "443" &&
-    apiBaseUrl.port !== ""
-  ) {
-    throw new Error("Agent API base URL must be HTTP or HTTPS");
-  }
-  if (apiBaseUrl.search !== "") {
-    throw new Error("Agent API base URL must not have a query string");
-  }
-  if (apiBaseUrl.hash !== "") {
-    throw new Error("Agent API base URL must not have a hash");
-  }
-
-  return {
-    id: agent.id,
-    onChainIdentifier: agent.onChainIdentifier,
-    name: agent.overrideName ?? agent.name,
-    description:
-      agent.overrideDescription ?? agent.description ?? t("defaultDescription"),
-    exampleOutput:
-      agent.overrideExampleOutput.length > 0
-        ? agent.overrideExampleOutput.map((example) => ({
-            id: example.id,
-            name: example.name,
-            mimeType: example.mimeType,
-            url: ipfsUrlResolver(example.url),
-          }))
-        : agent.exampleOutput.map((example) => ({
-            id: example.id,
-            name: example.name,
-            mimeType: example.mimeType,
-            url: ipfsUrlResolver(example.url),
-          })),
-    averageStars:
-      Number(agent.rating.totalRatings) === 0
-        ? null
-        : Math.min(
-            5,
-            Math.round(
-              Number(agent.rating.totalStars) /
-                Number(agent.rating.totalRatings),
-            ),
-          ),
-    author: agent.overrideAuthorName ?? agent.authorName,
-    legal: (() => {
-      const privacyPolicy =
-        agent.overrideLegalPrivacyPolicy ?? agent.legalPrivacyPolicy;
-      const terms = agent.overrideLegalTerms ?? agent.legalTerms;
-      const other = agent.overrideLegalOther ?? agent.legalOther;
-      return privacyPolicy || terms || other
-        ? { privacyPolicy, terms, other }
-        : undefined;
-    })(),
-    tags:
-      agent.overrideTags.length > 0
-        ? agent.overrideTags.map((tag) => tag.name)
-        : agent.tags.map((tag) => tag.name),
-    image: ipfsUrlResolver(agent.overrideImage ?? agent.image),
-    apiBaseUrl: apiBaseUrl,
-    credits: calculateCredits(
-      Number(agent.pricing.fixedPricing.amounts[0].amount),
-    ),
-  };
-}
